Apresentation Version

+++ editProdsPage.php
@@ -3,25 +3,25 @@
     
  require("admin/bd.php");
 if(isset($_POST["submit"])){
+
     $name = filter_var($_POST["name"], FILTER_SANITIZE_STRING);
-    $quantity = filter_var($_POST["quantity"], FILTER_SANITIZE_INT);
+    $quantity = filter_var($_POST["quantity"], FILTER_SANITIZE_NUMBER_INT);
     $country = filter_var($_POST["country"], FILTER_SANITIZE_STRING);
-    $price = filter_var($_POST["price"], FILTER_SANITIZE_INT);
+    $price = filter_var($_POST["price"], FILTER_SANITIZE_NUMBER_INT);
     $nota =filter_var($_POST["note"], FILTER_SANITIZE_STRING);
 
     $id = $_POST["userid"];
     $refe = $_POST["ref"];
     $ref = md5(time());
 
-    if (count($_FILES) > 0) {
-            
+    if (count($_FILES) > 0 && !empty($_FILES)) {
         $sql = "DELETE FROM image WHERE prodRef='" . $refe . "'";
           
-        $filepath = $_FILES['file']['tmp_name'];
+        $filepath = $_FILES['file']['tmp_name'][0];
         $fileSize = filesize($filepath);
 
-        if ($fileSize === 0) {
-            if ($fileSize > 3145728) {
+        if ($fileSize > 0) {
+            if ($fileSize < 3145728) {
             $countfiles = count($_FILES['file']['name']);
 
             if ($mysqli->query($sql) === TRUE) {
@@ -38,15 +38,12 @@ if(isset($_POST["submit"])){
                }
             }
             else {
-                echo "<meta http-equiv=refresh content='0; url=register.php?page=3&message=10'>";exit;
+                echo "<meta http-equiv=refresh content='0; url=index.php?page=3&message=11'>";exit;
             }
-        }
-        else {
-            echo "<meta http-equiv=refresh content='0; url=register.php?page=3&message=10'>";exit;
         }
     }
     else {
-        echo "<meta http-equiv=refresh content='0; url=register.php?page=3&message=10'>";exit;
+        echo "<meta http-equiv=refresh content='0; url=index.php?page=3&message=10'>";exit;
     }
     $sql = "UPDATE produtos SET nome='".$name."', preco='".$price."', quantidade='".$quantity."', pais='".$country."', nota='".$nota."' WHERE id='".$id."'";
             
@@ -70,7 +67,7 @@ if(isset($_POST["submit"])){
                 <div class="col-lg-12">
                     <div class="breadcrumb__links">
                         <a href="./index.html"><i class="fa fa-home"></i> Home</a>
-                        <span>Add Product</span>
+                        <span>Edit Product</span>
                     </div>
                 </div>
             </div>
